requirements.txt
flask
faiss-cpu
sentence-transformers
wikipedia
numpy
torch
transformers
scikit-learn
pandas
regex